% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cqq_scatter}
\alias{cqq_scatter}
\title{A scatter plot of two quantitative and one categorical variables}
\usage{
cqq_scatter(data, cat_1, quant_1, quant_2, title = "",
  x_label = quote(quant_1), y_label = quote(quant_2))
}
\arguments{
\item{data}{\code{(data.frame)} The data that will be plotted. Must have at least two numeric
columns and one factor column, preferably ordered.}

\item{quant_1}{(\code{character, length == 1}) The name of column in \code{data} that stores the
quantative that will plotted on the x-axis.}

\item{quant_2}{(\code{character, length == 1}) The name of column in \code{data} that stores the
quantative that will plotted on the y-axis.}

\item{title}{(\code{character, length == 1}) The tile of the graph}

\item{x_label}{(\code{character, length == 1}) The label of the x-axis}

\item{y_label}{(\code{character, length == 1}) The label of the y-axis}
}
\description{
A scatter plot that explores the relationship of two quantitave varaibles that are subdivded
into around 3-8 categories. Applies linear regression significance tests within categories and
on category means. 95% cofidence intervals of variable means are displayed on the plot.
}

