% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cqq_scatter_1}
\alias{cqq_scatter_1}
\title{A scatter plot of two quantitative and one categorical variables}
\usage{
cqq_scatter_1(data, cat_1, quant_1, quant_2, x_label = quant_1,
  y_label = quant_2, title = paste(quant_1, "vs", quant_2, "by", cat_1),
  text_size = 1)
}
\arguments{
\item{data}{\code{(data.frame)} The data that will be plotted. Must have at least two numeric
columns and one factor column, preferably ordered.}

\item{cat_1}{(\code{character, length == 1}) The name of column in \code{data} that stores the
categorical variable. This column should be a factor, preferably ordered, with 3 to 10 levels.}

\item{quant_1}{(\code{character, length == 1}) The name of column in \code{data} that stores the
quantative that will plotted on the x-axis.}

\item{quant_2}{(\code{character, length == 1}) The name of column in \code{data} that stores the
quantative that will plotted on the y-axis.}

\item{x_label}{(\code{character, length == 1}) The label of the x-axis}

\item{y_label}{(\code{character, length == 1}) The label of the y-axis}

\item{title}{(\code{character, length == 1}) The tile of the graph}

\item{text_size}{(\code{numeric, length == 1}) A multiplicative scaling of all text size. For,
example, 1 is standard size and 2 is twice as large. This is useful for exporting graphs of
various sizes, since text size does not automatically scale with export size.}
}
\description{
A scatter plot that explores the relationship of two quantitave variables that are subdivided
into 3-10 categories. Applies linear regression significance tests within categories and
on category means. 95\% bootstrapped cofidence intervals of variable means are displayed on the
plot.
}
\examples{
# Plot sepal width vs sepal length for three species of iris:
cqq_scatter_1(data = iris, cat_1 = "Species", quant_1 = "Sepal.Width", quant_2 = "Sepal.Length")
}

